feat: removed .gitignore and added restarter.sh
I'm so dummy that I put the main source code ignored. That is because the code was originally
written on my monorepo, and I just copied the stuff here to this public repository. The one that I
pushed on my monorepo is encrypted (as it contains my router password). That explains why
restarter.sh is encrypted. Make sense right?